task_user_id

+++ app/Http/Controllers/TasksController.php
@@ -6,18 +6,26 @@ use Illuminate\Http\Request;
 
 use App\Task;    // 追加
 
+use app\User;    // 追加
+
 class TasksController extends Controller
 {
+    public function __construct(){
+    $this->middleware('auth');
+  }
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::orderBy('id', 'desc')->paginate(10);
+        
+        /// メッセージ一覧を取得
+        $tasks = Task::where('user_id', \Auth::user()->id)->get();
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // メッセージ一覧ビューでそれを表示
+       
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -42,6 +50,7 @@ class TasksController extends Controller
         
         // タスクを作成
         $task = new Task;
+        $task->user_id = \Auth::id(); // 一行追加
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
@@ -55,11 +64,15 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-
         // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+           'task' => $task,
+            ]);
+        } else {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
     }
 
     // getでtasks/（任意のid）/editにアクセスされた場合の「更新画面表示処理」
@@ -67,11 +80,15 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-
         // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
-        ]);
+            ]);
+        } else {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
     }
 
     // putまたはpatchでtasks/（任意のid）にアクセスされた場合の「更新処理」
@@ -84,13 +101,13 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを更新
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
-
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、タスクを更新
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -98,11 +115,11 @@ class TasksController extends Controller
     // deleteでtasks/（任意のid）にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
